+++ .daruma.json
@@ -1,0 +1,4 @@
+{
+  "name": "daruma",
+  "isLibrary": false 
+}
