--- .eslintrc.js
@@ -1,4 +1,0 @@
-module.exports = {
-    "extends": "google",
-    "installedESLint": true
-};
