--- configs/daruma.json
@@ -1,4 +1,0 @@
-{
-  "name": "",
-  "isLibrary": false
-}
