--- configs/eslintrc.txt
@@ -1,6 +1,0 @@
-{
-    "extends": "airbnb",
-    "plugins": [
-        "react"
-    ]
-}
