--- helpers/createSpinner.js
@@ -1,7 +1,0 @@
-var ora = require('ora');
-var chalk = require('chalk');
-
-module.exports = function createSpinner(text) {
-  var spinner = ora(chalk.bold(text));
-  return spinner;
-};
