+++ package.json
@@ -4,7 +4,7 @@
   "description": "Build tool inspired by Rust's Cargo package manager",
   "main": "index.js",
   "scripts": {
-    "postinstall": "node helpers/postinstall.js",
+    "postinstall": "",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -29,7 +29,14 @@
     "vorpal": "^1.10.10"
   },
   "devDependencies": {
+    "babel-core": "^6.8.0",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-stage-0": "^6.5.0",
     "eslint": "^2.9.0",
-    "eslint-config-google": "^0.5.0"
+    "eslint-config-airbnb": "^8.0.0",
+    "eslint-plugin-import": "^1.6.1",
+    "eslint-plugin-jsx-a11y": "^1.0.4",
+    "eslint-plugin-promise": "^1.1.0",
+    "eslint-plugin-react": "^5.0.1"
   }
 }
