+++ src/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+    "extends": "airbnb",
+    "installedESLint": true,
+    "plugins": [
+        "react"
+    ]
+};
