+++ src/configs/daruma.json
@@ -1,0 +1,4 @@
+{
+  "name": "",
+  "isLibrary": false
+}
