+++ src/configs/gitignore.js
@@ -1,0 +1,37 @@
+export default {
+  content:
+  `# Logs
+  logs
+  *.log
+  npm-debug.log*
+
+  # Runtime data
+  pids
+  *.pid
+  *.seed
+
+  # Directory for instrumented libs generated by jscoverage/JSCover
+  lib-cov
+
+  # Coverage directory used by tools like istanbul
+  coverage
+
+  # Grunt intermediate storage (http://gruntjs.com/creating-plugins#storing-task-files)
+  .grunt
+
+  # node-waf configuration
+  .lock-wscript
+
+  # Compiled binary addons (http://nodejs.org/api/addons.html)
+  build/Release
+
+  # Dependency directory
+  node_modules
+
+  # Optional npm cache directory
+  .npm
+
+  # Optional REPL history
+  .node_repl_history
+  `,
+};
