+++ src/configs/webpack.config.js
@@ -1,0 +1,38 @@
+export default {
+  content:
+`var path = require('path');
+var libraryName = require('./.daruma.json').name;
+var outputFile = libraryName + '.js';
+
+var config = {
+  entry: __dirname + '/src/index.js',
+  devtool: 'source-map',
+  output: {
+    path: __dirname + '/dist',
+    filename: outputFile
+  },
+  module: {
+    loaders: [
+      {
+        test: /(\.jsx|\.js)$/,
+        loader: 'babel',
+        exclude: /(node_modules|bower_components)/,
+        query: {
+          presets: ['es2015', 'stage-0']
+        }
+      },
+      {
+        test: /\.json$/,
+        loader: 'json'
+      }
+    ]
+  },
+  resolve: {
+    root: path.resolve('./src'),
+    extensions: ['', '.js', '.json']
+  }
+};
+
+module.exports = config;
+`,
+};
